requirements.txt
flask==3.0.0
flask-cors==4.0.0
rembg==2.0.55
pillow==10.1.0
numpy==1.24.3
scipy==1.11.4
requests==2.31.0
torch==2.1.0
torchvision==0.16.0
diffusers==0.25.0
transformers==4.36.0
accelerate==0.25.0
controlnet-aux==0.0.7
opencv-python==4.8.1.78
clip-interrogator==0.6.0
xformers==0.0.23
